#545 improve inverse kinetics
