+++ docs/source/models/submodels/interface/inverse_kinetics/base_inverse_kinetics.rst
@@ -1,0 +1,5 @@
+Base Inverse Kinetics
+=====================
+
+.. autoclass:: pybamm.interface.inverse_kinetics.BaseInverseKinetics
+    :members:
