+++ docs/source/models/submodels/interface/inverse_kinetics/index.rst
@@ -5,4 +5,5 @@ Inverse Interface Kinetics
   :maxdepth: 1
 
   base_inverse_first_order_kinetics
+  base_inverse_kinetics
   inverse_butler_volmer
