+++ pybamm/models/submodels/interface/inverse_kinetics/__init__.py
@@ -1,5 +1,6 @@
+from .base_inverse_kinetics import BaseInverseKinetics
+from .base_inverse_first_order_kinetics import BaseInverseFirstOrderKinetics
 from .inverse_butler_volmer import (
     BaseInverseButlerVolmer,
     BaseInverseFirstOrderButlerVolmer,
 )
-from .base_inverse_first_order_kinetics import BaseInverseFirstOrderKinetics
