+++ pybamm/models/submodels/interface/inverse_kinetics/inverse_butler_volmer.py
@@ -4,12 +4,12 @@
 
 import pybamm
 import autograd.numpy as np
-from ..base_interface import BaseInterface
+from .base_inverse_kinetics import BaseInverseKinetics
 from .base_inverse_first_order_kinetics import BaseInverseFirstOrderKinetics
 from ..kinetics.butler_volmer import BaseButlerVolmer
 
 
-class BaseInverseButlerVolmer(BaseButlerVolmer):
+class BaseInverseButlerVolmer(BaseInverseKinetics, BaseButlerVolmer):
     """
     A base submodel that implements the inverted form of the Butler-Volmer relation to
     solve for the reaction overpotential.
@@ -22,63 +22,19 @@ class BaseInverseButlerVolmer(BaseButlerVolmer):
         The domain(s) in which to compute the interfacial current. Default is None,
         in which case j.domain is used.
 
+    **Extends:** :class:`pybamm.interface.kinetics.BaseButlerVolmer`
+
     """
 
     def __init__(self, param, domain):
         super().__init__(param, domain)
 
-    def get_coupled_variables(self, variables):
-        ocp, dUdT = self._get_open_circuit_potential(variables)
-
-        j0 = self._get_exchange_current_density(variables)
-        j_tot_av = self._get_average_total_interfacial_current_density(variables)
-        # Broadcast to match j0's domain
-        if j0.domain == []:
-            j = j_tot_av
-        else:
-            j = pybamm.Broadcast(
-                j_tot_av, [self.domain.lower() + " electrode"], broadcast_type="primary"
-            )
-
-        if self.domain == "Negative":
-            ne = self.param.ne_n
-        elif self.domain == "Positive":
-            ne = self.param.ne_p
-
-        eta_r = (2 / ne) * pybamm.Function(np.arcsinh, j / (2 * j0))
-        delta_phi = eta_r + ocp
-
-        variables.update(self._get_standard_interfacial_current_variables(j))
-        variables.update(
-            self._get_standard_total_interfacial_current_variables(j_tot_av)
-        )
-        variables.update(self._get_standard_exchange_current_variables(j0))
-        variables.update(self._get_standard_overpotential_variables(eta_r))
-        variables.update(
-            self._get_standard_surface_potential_difference_variables(delta_phi)
-        )
-        variables.update(self._get_standard_ocp_variables(ocp, dUdT))
-
-        if (
-            "Negative electrode" + self.reaction_name + " interfacial current density"
-            in variables
-            and "Positive electrode"
-            + self.reaction_name
-            + " interfacial current density"
-            in variables
-        ):
-            variables.update(
-                self._get_standard_whole_cell_interfacial_current_variables(variables)
-            )
-            variables.update(
-                self._get_standard_whole_cell_exchange_current_variables(variables)
-            )
-
-        return variables
+    def _get_inverse_kinetics(self, j, j0, ne):
+        return (2 / ne) * pybamm.Function(np.arcsinh, j / (2 * j0))
 
 
 class BaseInverseFirstOrderButlerVolmer(
-    BaseButlerVolmer, BaseInverseFirstOrderKinetics
+    BaseInverseFirstOrderKinetics, BaseButlerVolmer
 ):
     def __init__(self, param, domain):
         super().__init__(param, domain)
